autoscaler_predictor_model/requirements.txt
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.1.3
fonttools==4.44.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
numpy==1.26.2
packaging==23.2
pandas>=2.2.0
patsy==0.5.3
Pillow==10.1.0
pluggy==1.3.0
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn
scipy==1.11.3
six==1.16.0
statsmodels==0.14.0
threadpoolctl==3.2.0
tomli==2.0.1
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
apscheduler==3.10.1
requests==2.31.0
prometheus-client
kubernetes==25.3.0
xgboost
flask-swagger-ui
pyarrow
matplotlib
seaborn
holidays
prophet
fastapi
pydantic
uvicorn
python-multipart
aiohttp
